requirements.txt
PyAudio==0.2.14
faster_whisper==1.1.1
librosa==0.10.2
soundfile==0.13.1
fastapi==0.115.11 
ffmpeg-python==0.2.0
uvicorn[standard]==0.34.0
silero-vad==5.1.2
vllm==0.7.3
